+++ src/main/java/com/tracer/exception/ExpiredTokenException.java
@@ -1,0 +1,2 @@
+package com.tracer.exception;public class ExpiredTokenException {
+}
