+++ src/main/java/com/tracer/repository/tokens/TokenRepository.java
@@ -1,0 +1,2 @@
+package com.tracer.repository.tokens;public class TokenRepository {
+}
