+++ src/main/java/com/tracer/util/JwtUtils.java
@@ -1,0 +1,2 @@
+package com.tracer.util;public class JwtUtils {
+}
